# api/openapi.yaml
openapi: "3.1.0"
info:
  title: FamilyRoots API
  version: "1.0.0"
  description: |
    ## FamilyRoots REST API — v1

    A multi-tenant genealogy platform API following REST conventions.

    ### Versioning
    URL-path versioning: all v1 endpoints are prefixed `/api/v1`.
    Breaking changes increment the major version (`/api/v2`).
    Minor/patch changes are backward-compatible and do not change the URL.

    ### Authentication
    Two tokens are issued on login:
    - **Access token** — short-lived JWT (15 min), passed as `Authorization: Bearer <token>`
    - **Refresh token** — long-lived (30 days), stored as an `HttpOnly SameSite=Strict` cookie

    ### Multi-tenancy
    Every request is scoped to a tenant derived from the JWT `tenant_id` claim.
    Tenant is never supplied by the caller — it is always resolved server-side from the token.

    ### Rate limiting
    All endpoints return rate-limit headers:
    - `X-RateLimit-Limit` — requests allowed in the current window
    - `X-RateLimit-Remaining` — requests remaining in the current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    - `Retry-After` — seconds to wait (only present on 429 responses)

    Default limits by plan:
    | Plan        | Requests/min | Burst |
    |-------------|-------------|-------|
    | FREE        | 60          | 10    |
    | BASIC       | 300         | 30    |
    | PREMIUM     | 1000        | 100   |
    | PROFESSIONAL| 5000        | 500   |

...
    ```
...
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued on login. Expires in 15 minutes.
        Payload includes: sub (user_id), tenant_id, email, roles, plan.

    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        HttpOnly SameSite=Strict cookie holding the refresh token.
        Used only on POST /auth/refresh and POST /auth/logout.

  # ──────────────────────────────────────────────────────────
  # COMMON PARAMETERS
  # ──────────────────────────────────────────────────────────
  parameters:
    treeId:
      name: treeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Family tree identifier
...
  schemas:
...
    CollaboratorRole:
      type: string
      enum: [VIEWER, CONTRIBUTOR, EDITOR, ADMIN]

    # ── Error ───────────────────────────────────────────────
    Error:
      type: object
      required: [type, title, status, requestId]
      properties:
        type:
          type: string
          format: uri
          example: "https://familyroots.app/errors/validation-error"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: "The request body contains invalid fields."
        instance:
          type: string
          example: "/api/v1/trees/abc/persons"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        requestId:
          type: string
          example: "req_01hzab3k7x"
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
          example: "sex"
        code:
          type: string
          example: "INVALID_ENUM"
        message:
          type: string
          example: "Must be one of: MALE, FEMALE, OTHER, UNKNOWN"

    # ── Pagination ──────────────────────────────────────────
...
    TokenPair:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
          description: Short-lived JWT (15 min)
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          example: 900
          description: Access token TTL in seconds
        user:
          $ref: "#/components/schemas/UserProfile"

    # ── User ────────────────────────────────────────────────
    UserProfile:
      type: object
      required: [id, email, displayName, isVerified, plan, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          maxLength: 100
        avatarUrl:
          type: string
          format: uri
          nullable: true
        locale:
          type: string
          example: "en"
        timezone:
          type: string
          example: "America/New_York"
        isVerified:
          type: boolean
        plan:
          type: string
          enum: [FREE, BASIC, PREMIUM, FAMILY, PROFESSIONAL]
        createdAt:
          type: string
          format: date-time
...
    Invitation:
      type: object
      required: [id, email, role, status, expiresAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/CollaboratorRole"
        status:
          type: string
          enum: [PENDING, ACCEPTED, DECLINED, EXPIRED, REVOKED]
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    # ── Person ──────────────────────────────────────────────
...
  responses:
    400BadRequest:
      description: Bad Request — malformed JSON or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    401Unauthorized:
      description: Unauthorized — missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    403Forbidden:
      description: Forbidden — valid token but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    422UnprocessableEntity:
      description: Unprocessable Entity — validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    429TooManyRequests:
      description: Too Many Requests — rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

# ──────────────────────────────────────────────────────────────
# GLOBAL SECURITY (override per-operation where needed)
# ──────────────────────────────────────────────────────────────
security:
  - BearerAuth: []

# ──────────────────────────────────────────────────────────────
# TAGS
# ──────────────────────────────────────────────────────────────
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      operationId: register
      summary: Register a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, displayName]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                password:
                  type: string
                  minLength: 10
                  maxLength: 128
                  description: >
                    Must contain: uppercase, lowercase, digit, special character.
                    Must not appear in HIBP breach database.
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 100
                locale:
                  type: string
                  default: "en"
                timezone:
                  type: string
                  default: "UTC"
      responses:
        "201":
          description: Account created — verification email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/422UnprocessableEntity"
        "429":
          $ref: "#/components/responses/429TooManyRequests"
  /auth/login:
    post:
      tags: [Authentication]
      operationId: login
      summary: Authenticate with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                rememberMe:
                  type: boolean
                  default: false
                  description: Extends refresh token TTL from 30 days to 90 days
      responses:
        "200":
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
              description: "HttpOnly SameSite=Strict refresh_token cookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "423":
          description: Account locked after too many failed attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/429TooManyRequests"
  /auth/logout:
    post:
      tags: [Authentication]
      operationId: logout
      summary: Invalidate the current session
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        "204":
          description: Logged out — refresh token revoked and cookie cleared
        "401":
          $ref: "#/components/responses/401Unauthorized"
  /auth/refresh:
    post:
      tags: [Authentication]
      operationId: refreshToken
      summary: Issue a new access token using the refresh cookie
      security:
        - CookieAuth: []
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, expiresIn]
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: integer
                    example: 900
        "401":
          description: Refresh token missing, expired, or revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/429TooManyRequests"
  /auth/verify-email:
    post:
      tags: [Authentication]
      operationId: verifyEmail
      summary: Verify email address with token from verification email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  minLength: 64
                  maxLength: 64
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          description: Token invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/forgot-password:
    post:
      tags: [Authentication]
      operationId: forgotPassword
      summary: Send a password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "202":
          description: Reset email sent if account exists (always returns 202 to prevent enumeration)
        "429":
          $ref: "#/components/responses/429TooManyRequests"
  /auth/reset-password:
    post:
      tags: [Authentication]
      operationId: resetPassword
      summary: Reset password using token from reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  minLength: 10
                  maxLength: 128
      responses:
        "200":
          description: Password reset — new token pair issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          description: Token invalid, expired, or password too weak
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /trees/{treeId}/invitations:
    parameters:
      - $ref: "#/components/parameters/treeId"
    get:
      tags: [Trees]
      operationId: listInvitations
      summary: List pending invitations for a tree
      responses:
        "200":
          description: Invitations list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invitation"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"

    post:
      tags: [Trees]
      operationId: sendInvitation
      summary: Send an invitation to collaborate on a tree
      description: ADMIN or owner only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email:
                  type: string
                  format: email
                role:
                  $ref: "#/components/schemas/CollaboratorRole"
      responses:
        "201":
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "403":
          $ref: "#/components/responses/403Forbidden"
        "409":
          description: Invitation already pending for this email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...